+++ src/utils/componentCategories.ts
@@ -84,6 +84,7 @@ const categories = {
     components: [
       "useCarouselStyles",
       "useCloseDropDown",
+      "useFilterBar",
       "useMultiSelect",
       "useNYPLBreakpoints",
       "useNYPLTheme",